refactor: 멀티 모듈 실행 및 인수 테스트를 위한 모듈 생성
