+++ configuration/src/main/java/com/ticketlounge/TicketLoungeApplication.java
@@ -1,13 +1,13 @@
-package com.ticketexchange;
+package com.ticketlounge;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class TicketExchangeApplication {
+public class TicketLoungeApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(TicketExchangeApplication.class, args);
+        SpringApplication.run(TicketLoungeApplication.class, args);
     }
 
 }
