+++ configuration/src/main/resources/db/migration/V1.0__create_table.sql
