+++ configuration/src/test/java/com/ticketlounge/AcceptanceTest/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package com.ticketexchange.common.acceptance;
+package com.ticketlounge.AcceptanceTest;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,11 +6,15 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
+import org.springframework.test.context.ContextConfiguration;
+
+import com.ticketlounge.TestApplication;
 
 import io.restassured.RestAssured;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
+@ContextConfiguration(classes = TestApplication.class)
 public abstract class AcceptanceTest {
 
     @LocalServerPort
