+++ configuration/src/test/java/com/ticketlounge/TestApplication.java
@@ -1,0 +1,7 @@
+package com.ticketlounge;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestApplication {
+}
