+++ configuration/src/test/java/com/ticketlounge/member/api/MemberApiSupporter.java
@@ -1,11 +1,10 @@
-package com.ticketexchange.member.api;
+package com.ticketlounge.member.api;
 
-import static com.ticketexchange.member.fixture.MemberFixture.로그인_요청;
-import static com.ticketexchange.member.fixture.MemberFixture.회원가입_요청;
+import static com.ticketlounge.member.fixture.MemberFixture.로그인_요청;
+import static com.ticketlounge.member.fixture.MemberFixture.회원가입_요청;
 import static io.restassured.RestAssured.given;
 
-import org.springframework.http.MediaType;
-
+import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 
@@ -17,8 +16,8 @@ public final class MemberApiSupporter {
     public static ExtractableResponse<Response> 회원가입() {
         return given()
                 .body(회원가입_요청())
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .contentType(ContentType.JSON)
+                .accept(ContentType.JSON)
                 .when()
                 .post("/api/v1/members")
                 .then()
@@ -28,8 +27,8 @@ public final class MemberApiSupporter {
     public static ExtractableResponse<Response> 로그인() {
         return given()
                 .body(로그인_요청())
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .contentType(ContentType.JSON)
+                .accept(ContentType.JSON)
                 .when()
                 .post("/api/v1/login")
                 .then()
